Extract AutoCompleteUtil
