+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/ConversationViewModel.kt
@@ -49,9 +49,10 @@ import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ser
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.chatlist.ConversationChatListUseCase
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.post.post_actions.PostActionFlags
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.InitialReplyTextProvider
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompleteCategory
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompleteHint
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompleteHintCollection
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompleteIcon
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompleteType
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.ReplyAutoCompleteHintProvider
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.thread.ConversationThreadUseCase
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.MetisContext
@@ -586,7 +587,7 @@ internal open class ConversationViewModel(
     override fun produceAutoCompleteHints(
         tagChar: Char,
         query: String
-    ): Flow<DataState<List<AutoCompleteCategory>>> = when (tagChar) {
+    ): Flow<DataState<List<AutoCompleteHintCollection>>> = when (tagChar) {
         '@' -> {
             produceUserMentionAutoCompleteHints(query)
         }
@@ -611,7 +612,7 @@ internal open class ConversationViewModel(
             }
         }
 
-    private fun produceUserMentionAutoCompleteHints(query: String): Flow<DataState<List<AutoCompleteCategory>>> =
+    private fun produceUserMentionAutoCompleteHints(query: String): Flow<DataState<List<AutoCompleteHintCollection>>> =
         flatMapLatest(
             accountService.authToken,
             serverConfigurationService.serverUrl
@@ -625,8 +626,8 @@ internal open class ConversationViewModel(
                         serverUrl = serverUrl
                     )
                     .bind { users ->
-                        AutoCompleteCategory(
-                            name = R.string.markdown_textfield_autocomplete_category_users,
+                        AutoCompleteHintCollection(
+                            type = AutoCompleteType.USERS,
                             items = users.map {
                                 AutoCompleteHint(
                                     it.name.orEmpty(),
@@ -640,7 +641,7 @@ internal open class ConversationViewModel(
             }
         }
 
-    private fun produceExerciseAndLectureAutoCompleteHints(query: String): Flow<DataState<List<AutoCompleteCategory>>> =
+    private fun produceExerciseAndLectureAutoCompleteHints(query: String): Flow<DataState<List<AutoCompleteHintCollection>>> =
         course.map { courseDataState ->
             courseDataState.bind { course ->
                 val exerciseAutoCompleteItems =
@@ -680,19 +681,19 @@ internal open class ConversationViewModel(
                         }
 
                 listOf(
-                    AutoCompleteCategory(
-                        name = R.string.markdown_textfield_autocomplete_category_exercises,
+                    AutoCompleteHintCollection(
+                        type = AutoCompleteType.EXERCISES,
                         items = exerciseAutoCompleteItems
                     ),
-                    AutoCompleteCategory(
-                        name = R.string.markdown_textfield_autocomplete_category_lectures,
+                    AutoCompleteHintCollection(
+                        type = AutoCompleteType.LECTURES,
                         items = lectureAutoCompleteItems
                     )
                 )
             }
         }
 
-    private fun produceConversationAutoCompleteHints(query: String): Flow<DataState<List<AutoCompleteCategory>>> =
+    private fun produceConversationAutoCompleteHints(query: String): Flow<DataState<List<AutoCompleteHintCollection>>> =
         conversations.map { conversationsDataState ->
             conversationsDataState.bind { conversations ->
                 val conversationAutoCompleteItems = conversations
@@ -708,8 +709,8 @@ internal open class ConversationViewModel(
                     }
 
                 listOf(
-                    AutoCompleteCategory(
-                        name = R.string.markdown_textfield_autocomplete_category_channels,
+                    AutoCompleteHintCollection(
+                        type = AutoCompleteType.CHANNELS,
                         items = conversationAutoCompleteItems
                     )
                 )
