+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/MarkdownTextField.kt
@@ -62,6 +62,7 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.PopupProperties
 import de.tum.informatics.www1.artemis.native_app.core.ui.markdown.MarkdownText
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.R
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompleteType
 import kotlinx.coroutines.launch
 
 
@@ -83,7 +84,7 @@ internal fun MarkdownTextField(
     onFocusAcquired: () -> Unit = {},
     onFocusLost: () -> Unit = {},
     onTextChanged: (TextFieldValue) -> Unit,
-    showAutoCompletePopup: ((AutocompleteType) -> Unit)? = null,
+    showAutoCompletePopup: ((AutoCompleteType) -> Unit)? = null,
     formattingOptionButtons: @Composable () -> Unit = {},
 ) {
     val text = textFieldValue.text
@@ -231,7 +232,7 @@ private fun TextFieldOptions(
     isPreviewEnabled: Boolean,
     showFormattingOptions: Boolean,
     onChangeViewType: (ViewType) -> Unit,
-    showAutoCompletePopup: ((AutocompleteType) -> Unit)? = null,
+    showAutoCompletePopup: ((AutoCompleteType) -> Unit)? = null,
     formattingOptionButtons: @Composable () -> Unit = {},
     onOpenFilePicker: () -> Unit = {},
     onOpenImagePicker: () -> Unit = {}
@@ -365,7 +366,7 @@ private fun TextFieldOptionsIconButton(
 @Composable
 private fun TaggingDropdownMenu(
     isTaggingDropdownExpanded: Boolean,
-    showAutoCompletePopup: ((AutocompleteType) -> Unit)?,
+    showAutoCompletePopup: ((AutoCompleteType) -> Unit)?,
     onDismissRequest: () -> Unit,
 ) {
     DropdownMenu(
@@ -385,7 +386,7 @@ private fun TaggingDropdownMenu(
             text = { Text(text = stringResource(R.string.reply_format_mention_members)) },
             onClick = {
                 onDismissRequest()
-                showAutoCompletePopup?.invoke(AutocompleteType.USERS)
+                showAutoCompletePopup?.invoke(AutoCompleteType.USERS)
             }
         )
 
@@ -399,7 +400,7 @@ private fun TaggingDropdownMenu(
             text = { Text(stringResource(R.string.reply_format_mention_channels)) },
             onClick = {
                 onDismissRequest()
-                showAutoCompletePopup?.invoke(AutocompleteType.CHANNELS)
+                showAutoCompletePopup?.invoke(AutoCompleteType.CHANNELS)
             }
         )
 
@@ -413,7 +414,7 @@ private fun TaggingDropdownMenu(
             text = { Text(stringResource(R.string.reply_format_mention_exercises)) },
             onClick = {
                 onDismissRequest()
-                showAutoCompletePopup?.invoke(AutocompleteType.EXERCISES)
+                showAutoCompletePopup?.invoke(AutoCompleteType.EXERCISES)
             }
         )
 
@@ -427,7 +428,7 @@ private fun TaggingDropdownMenu(
             text = { Text(stringResource(R.string.reply_format_mention_lectures)) },
             onClick = {
                 onDismissRequest()
-                showAutoCompletePopup?.invoke(AutocompleteType.LECTURES)
+                showAutoCompletePopup?.invoke(AutoCompleteType.LECTURES)
             }
         )
     }
