+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/ReplyTextField.kt
@@ -5,7 +5,6 @@ import android.net.Uri
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.annotation.StringRes
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
@@ -31,7 +30,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
@@ -50,31 +48,29 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.text.input.getTextBeforeSelection
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat.getString
-import de.tum.informatics.www1.artemis.native_app.core.data.DataState
 import de.tum.informatics.www1.artemis.native_app.core.ui.Spacings
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.R
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.MetisModificationFailure
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.model.FileValidationConstants
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompleteCategory
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.AutoCompleteType
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.LocalReplyAutoCompleteHintProvider
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.ReplyAutoCompletePopup
-import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.ReplyAutoCompletePopupPositionProvider
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.util.AutoCompleteUtil
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.util.AutoCompleteUtil.manageAutoCompleteHints
+import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.util.ReplyAutoCompletePopupPositionProvider
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.util.MarkdownListContinuationUtil.continueListIfApplicable
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.util.MarkdownStyleUtil
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.Flow
 
 internal const val TEST_TAG_CAN_CREATE_REPLY = "TEST_TAG_CAN_CREATE_REPLY"
 internal const val TEST_TAG_REPLY_TEXT_FIELD = "TEST_TAG_REPLY_TEXT_FIELD"
@@ -91,7 +87,7 @@ internal fun ReplyTextField(
     conversationName: String
 ) {
     val replyState: ReplyState = rememberReplyState(replyMode, updateFailureState)
-    val requestedAutoCompleteType = remember { mutableStateOf<AutocompleteType?>(null) }
+    val requestedAutoCompleteType = remember { mutableStateOf<AutoCompleteType?>(null) }
     var boxWidth by remember { mutableIntStateOf(0) }
     var popupMaxHeight by remember { mutableIntStateOf(0) }
 
@@ -178,7 +174,7 @@ internal fun ReplyTextField(
 @Composable
 private fun AutoCompletionDialog(
     replyMode: ReplyMode,
-    requestedAutoCompleteType: MutableState<AutocompleteType?>,
+    requestedAutoCompleteType: MutableState<AutoCompleteType?>,
     targetWidth: Dp,
     maxHeightFromScreen: Dp,
 ) {
@@ -216,7 +212,7 @@ private fun AutoCompletionDialog(
         maxHeightFromScreen = maxHeightFromScreen,
         popupPositionProvider = ReplyAutoCompletePopupPositionProvider,
         performAutoComplete = { replacement ->
-            val newTextFieldValue = performAutoComplete(
+            val newTextFieldValue = AutoCompleteUtil.perform(
                 currentTextFieldValue,
                 tagChars,
                 replacement
@@ -261,7 +257,7 @@ private fun CreateReplyUi(
     focusRequester: FocusRequester = remember { FocusRequester() },
     onReply: () -> Unit,
     onFileSelected: (Uri) -> Unit,
-    onRequestAutocompleteType: (AutocompleteType) -> Unit
+    onRequestAutocompleteType: (AutoCompleteType) -> Unit
 ) {
     var prevReplyContent by remember { mutableStateOf("") }
     var displayTextField: Boolean by remember { mutableStateOf(false) }
@@ -331,10 +327,10 @@ private fun CreateReplyUi(
             onRequestAutocompleteType(it)
             val newTextFieldValue = MarkdownStyleUtil.apply(
                 style = when (it) {
-                    AutocompleteType.USERS -> MarkdownStyle.UserMention
-                    AutocompleteType.CHANNELS -> MarkdownStyle.ChannelMention
-                    AutocompleteType.LECTURES -> MarkdownStyle.LectureMention
-                    AutocompleteType.EXERCISES -> MarkdownStyle.ExerciseMention
+                    AutoCompleteType.USERS -> MarkdownStyle.UserMention
+                    AutoCompleteType.CHANNELS -> MarkdownStyle.ChannelMention
+                    AutoCompleteType.LECTURES -> MarkdownStyle.LectureMention
+                    AutoCompleteType.EXERCISES -> MarkdownStyle.ExerciseMention
                 },
                 currentTextFieldValue = currentTextFieldValue,
             )
@@ -435,109 +431,6 @@ private fun SendButton(
     }
 }
 
-private fun performAutoComplete(
-    textFieldValue: TextFieldValue,
-    tagChars: List<Char>,
-    replacement: String
-): TextFieldValue {
-    // Perform replace in text
-    val replacementStart = textFieldValue
-        .getTextBeforeSelection(Int.MAX_VALUE)
-        .indexOfLastWhileTag(tagChars)
-        ?: return textFieldValue
-
-    val replacementEnd = textFieldValue.selection.min
-
-    val newText = textFieldValue
-        .text
-        .replaceRange(replacementStart, replacementEnd, replacement)
-
-    return textFieldValue.copy(
-        text = newText,
-        // Put cursor after replacement.
-        selection = TextRange(replacementStart + replacement.length)
-    )
-}
-
-/**
- * @return a list of auto complete hints that should be displayed, or null if no auto complete hints are to be displayed.
- */
-@Composable
-private fun manageAutoCompleteHints(
-    textFieldValue: TextFieldValue,
-    requestedAutocompleteType: AutocompleteType? = null
-): List<AutoCompleteCategory>? {
-    val replyAutoCompleteHintProvider = LocalReplyAutoCompleteHintProvider.current
-    var replyAutoCompleteHintProducer: Flow<DataState<List<AutoCompleteCategory>>>? by remember {
-        mutableStateOf(
-            null
-        )
-    }
-
-    val tagChars = replyAutoCompleteHintProvider.legalTagChars
-    LaunchedEffect(textFieldValue, replyAutoCompleteHintProvider) {
-        // Check that no text is selected, and instead we have a simple cursor
-        replyAutoCompleteHintProducer =
-            textFieldValue.getAutoCompleteReplacementTextFirstIndex(tagChars)?.let { tagIndex ->
-                val tagChar = textFieldValue.text[tagIndex]
-                val replacementWord = if (textFieldValue.text.length > tagIndex + 1) {
-                    textFieldValue.text.substring(tagIndex + 1).takeWhileTag()
-                } else ""
-
-                replyAutoCompleteHintProvider.produceAutoCompleteHints(tagChar, replacementWord)
-            }
-    }
-
-    val replyAutoCompleteHints = replyAutoCompleteHintProducer
-        ?.collectAsState(DataState.Loading())
-        ?.value
-        ?.orElse(emptyList())
-        .orEmpty()
-        .toMutableList()
-        .apply {
-            requestedAutocompleteType?.let { type ->
-                retainAll { it.name == type.title }
-            }
-        }
-
-    var latestValidAutoCompleteHints: List<AutoCompleteCategory>? by remember {
-        mutableStateOf(null)
-    }
-
-    LaunchedEffect(replyAutoCompleteHintProducer, replyAutoCompleteHints) {
-        if (replyAutoCompleteHintProducer == null) {
-            latestValidAutoCompleteHints = null
-        } else if (replyAutoCompleteHints.isNotEmpty()) {
-            latestValidAutoCompleteHints = replyAutoCompleteHints
-        }
-    }
-
-    return latestValidAutoCompleteHints
-}
-
-enum class AutocompleteType(@StringRes val title: Int) {
-    USERS(R.string.markdown_textfield_autocomplete_category_users),
-    CHANNELS(R.string.markdown_textfield_autocomplete_category_channels),
-    LECTURES(R.string.markdown_textfield_autocomplete_category_lectures),
-    EXERCISES(R.string.markdown_textfield_autocomplete_category_exercises),
-}
-
-/**
- * @return the index of the tagging char of the current replacement text, or null if the user is currently
- * not entering a replaceable text.
- */
-private fun TextFieldValue.getAutoCompleteReplacementTextFirstIndex(tagChars: List<Char>): Int? {
-    return if (selection.collapsed) {
-        // Gather the last word, meaning the characters from the last whitespace until the cursor.
-        val tagCharIndex = getTextBeforeSelection(Int.MAX_VALUE)
-            .indexOfLastWhileTag(tagChars)
-            ?: return null
-
-        val tagChar = text[tagCharIndex]
-
-        if (tagChar in tagChars) tagCharIndex else null
-    } else null
-}
 
 @Composable
 @Preview
@@ -560,41 +453,4 @@ private fun ReplyTextFieldPreview() {
             onFileSelected = { _ -> }
         )
     }
-}
-
-private fun CharSequence.indexOfLastWhileTag(tagChars: List<Char>): Int? {
-    var foundWhitespace = false
-
-    for (index in lastIndex downTo 0) {
-        val currentChar = this[index]
-
-        if (currentChar.isWhitespace() && foundWhitespace) {
-            return index + 1
-        } else if (currentChar in tagChars) {
-            return index
-        } else if (currentChar.isWhitespace()) {
-            foundWhitespace = true
-        }
-    }
-
-    return if (isEmpty()) null else 0
-}
-
-/**
- * Takes characters until the end of the string or a second whitespace has been found
- */
-private fun String.takeWhileTag(): String {
-    var foundWhitespace = false
-
-    for (index in indices) {
-        val currentChar = this[index]
-
-        if (currentChar.isWhitespace() && foundWhitespace) {
-            return substring(0, index)
-        } else if (currentChar.isWhitespace()) {
-            foundWhitespace = true
-        }
-    }
-
-    return this
 }
