--- feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/autocomplete/AutoCompleteCategory.kt
@@ -1,5 +1,0 @@
-package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete
-
-import androidx.annotation.StringRes
-
-class AutoCompleteCategory(@StringRes val name: Int, val items: List<AutoCompleteHint>)
