+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/autocomplete/AutoCompleteHintCollection.kt
@@ -1,0 +1,6 @@
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete
+
+data class AutoCompleteHintCollection(
+    val type: AutoCompleteType,
+    val items: List<AutoCompleteHint>
+)
