+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/autocomplete/util/ReplyAutoCompletePopupPositionProvider.kt
@@ -1,4 +1,4 @@
-package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete.util
 
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.IntRect
